# packages/backend/src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Matrix Flow API",
    "version": "1.0.0",
    "description": "API documentation for Matrix Flow application",
    "contact": {
      "name": "API Support",
      "url": "https://github.com/yourusername/matrix-flow"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3004",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/project": {
      "get": {
        "tags": [
          "Project"
        ],
        "summary": "Get all projects",
        "responses": {
          "200": {
            "description": "List of all projects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Project"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Project"
        ],
        "summary": "Create a new project",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project created successfully"
          }
        }
      }
    },
    "/api/project/{id}": {
      "get": {
        "tags": [
          "Project"
        ],
        "summary": "Get a project by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project details"
          }
        }
      },
      "put": {
        "tags": [
          "Project"
        ],
        "summary": "Update a project",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Project"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project updated successfully"
          }
        }
      },
      "delete": {
        "tags": [
          "Project"
        ],
        "summary": "Delete a project",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project deleted successfully"
          }
        }
      }
    },
    "/api/trigger": {
      "get": {
        "tags": [
          "Trigger"
        ],
        "summary": "Get all triggers",
        "responses": {
          "200": {
            "description": "List of all triggers"
          }
        }
      },
      "post": {
        "tags": [
          "Trigger"
        ],
        "summary": "Create a new trigger",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "nodeId",
                  "type",
                  "name"
                ],
                "properties": {
                  "nodeId": {
                    "type": "integer"
                  },
                  "type": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "config": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Trigger created successfully"
          }
        }
      }
    },
    "/api/trigger/{id}": {
      "get": {
        "tags": [
          "Trigger"
        ],
        "summary": "Get a trigger by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trigger details"
          }
        }
      },
      "put": {
        "tags": [
          "Trigger"
        ],
        "summary": "Update a trigger",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "config": {
                    "type": "object"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "inactive"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Trigger updated successfully"
          }
        }
      },
      "delete": {
        "tags": [
          "Trigger"
        ],
        "summary": "Delete a trigger",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trigger deleted successfully"
          }
        }
      }
    }
  },
...
}